Allow completion functions to associate data with completion items
When using v:completed_item, it is useful to be able to perform
subsequent actions based on the completed item. This allows user defined
completion functions to supply custom data (in the form of a string, for
simplicity), which can be used within the CompleteDone autocommand to
perform any subsequent actions.

Example use cases include:
 - automatic insertion of import or include statements
 - insertion of snippets for method arguments.

Fixes #2508