Improve expression evaluation when debug_backtrace_level is set
This change starts to populate the exestack with some data which can be
used to construct a more complete stack trace, in particular the real
file/line number and script contexts of any frame.

The first usage of this is to improve the expression evaluation in
debugger. Specifically this change uses the exestack to func the
funccal and script item for the frame currently being introspected, as
opposed to whatver happens to be executing, allowing more precise and
predictable introspection.

The tess are not comprehensive however curently. More are required to
cover:

- lambdas
- autocommands
- modelines
- etc.