Basic first steps for vim9script debugging
Upstream added some support for debug instructions in vim9script, which
mostly just-works at the top level for us. But in order to get the right
stack frame, we decide to store the dfunc_T in the estack (when
executing a def funciton) and the ufunc_T only when compiling it. This
gives us more access to things like the locals aund the instructions and
the stack (though we dcon't currently use them).

You can step and inspec local variables in the watch window, but
debugging is still quite limited. More todo:

- variables window (add locals, and script variables)
- use the user-freindly function name in the call stack (currently has
  the magic string rather than <sid> type stuff)
- lots more testing
- refactor so that we don't have to put dfunc_T in structs.h
- probably more, this was just a first pass